chore: restructure scan mod
